userMapper er færdig test

+++ src/main/java/app/entities/User.java
@@ -5,7 +5,7 @@ public class User {
     private String password;
     private double balance;
 
-    public User(String email, String password, int balance) {
+    public User(String email, String password, double balance) {
         this.email = email;
         this.password = password;
         this.balance = balance;
